Separate Byte Reader types and functions from test data buffer processor.h
