+++ src/Public/Framework/TestDataBufferProcessor.h
@@ -4,6 +4,7 @@
 
 #include "Framework/HLSLTypes.h"
 #include "Framework/TestDataBufferLayout.h"
+#include "Framework/TypeByteReader.h"
 #include "Utility/Concepts.h"
 #include "Utility/OverloadSet.h"
 #include "Utility/Type.h"
@@ -33,13 +34,6 @@ namespace stf
         u32 NumBytesStringData = 0;
         u32 NumSections = 0;
     };
-
-    using SingleTypeByteReader = std::function<std::string(std::span<const std::byte>)>;
-    using MultiTypeByteReader = std::function<std::string(const u16, std::span<const std::byte>)>;
-
-    std::string DefaultByteReader(const u16, std::span<const std::byte> InBytes);
-
-    using MultiTypeByteReaderMap = std::vector<MultiTypeByteReader>;
 
     struct FailedAssert
     {
@@ -110,38 +104,4 @@ namespace stf
         const TestDataBufferLayout& InLayout,
         std::span<const std::byte> InTestData,
         const MultiTypeByteReaderMap& InByteReaderMap);
-
-    template<HLSLTypeTriviallyConvertibleType T>
-    SingleTypeByteReader CreateSingleTypeReader()
-    {
-        return
-            [](const std::span<const std::byte> InBytes) -> std::string
-            {
-                if (InBytes.size_bytes() != sizeof(T))
-                {
-                    return std::format("Unexpected num bytes: {} for type {}", InBytes.size_bytes(), TypeToString<T>());
-                }
-
-                T data;
-                std::memcpy(&data, InBytes.data(), sizeof(T));
-
-                return std::format("{}", data);
-            };
-    }
-
-    template<HLSLTypeTriviallyConvertibleType... T>
-    MultiTypeByteReader CreateMultiTypeByteReader()
-    {
-        static constexpr auto numTypes = sizeof...(T);
-        return
-            [readers = std::array{ (CreateSingleTypeReader<T>())...}](const u16 InReaderId, const std::span<const std::byte> InBytes) -> std::string
-            {
-                if (InReaderId >= readers.size())
-                {
-                    return std::format("ReaderId must be less than the number of types used to construct this byte reader. ReaderId = {}, NumTypes = {}", InReaderId, numTypes);
-                }
-
-                return readers[InReaderId](InBytes);
-            };
-    }
 }
